pyproject.toml
[tool.poetry]
name = "gentoogram-bot"
version = "0.1.0"
description = ""
authors = ["Lee Watson <rev@revthefox.co.uk>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^12.0.0"
sentry-sdk = "^0.14.0"
pyyaml = "^5.3"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
